Move dump-string parsing to FreeCellGameService
Refactored FromDumpString and related parsing logic from FreeCellGameBase to FreeCellGameService, ensuring deserialization produces a service instance. Updated all usages and tests to use the new location. Added a new solver test for loading and solving from a dump string. FreeCellGameBase is now a more abstract base class.
